Change "name" to "region" and move load shp to utils
- The load shapefile function makes more sense in utils than generators.
- Change IPM_Region column to region and require the same name in user
geo files.
